importScripts() should throw exception with a reason when it fails to load a script from the network https://bugs.webkit.org/show_bug.cgi?id=184345
Reviewed by Alexey Proskuryakov.

Source/WebCore:

Currently importScripts() throws a NetworkError exception without a reason when
it fails to load a script from the network. It is ambiguous whether the load failed
due to an access control failure or some other error. Instead we should include
the reason the load failed in the thrown exception.

* workers/WorkerGlobalScope.cpp:
(WebCore::WorkerGlobalScope::importScripts):

LayoutTests:

Update expected results.

* http/tests/security/contentSecurityPolicy/worker-blob-inherits-csp-importScripts-redirect-cross-origin-blocked-expected.txt:
* http/tests/security/contentSecurityPolicy/worker-csp-importScripts-redirect-cross-origin-blocked-expected.txt:
* http/tests/workers/worker-importScripts-expected.txt:

Identifier: 199921@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@230345 268f45cc-cd09-0410-ab3c-d52691b4dbfc